+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 * Update `jackson-jr` to v2.9.5 to fix an incompatibility with Java 9
+* Add support for `SubscriptionDetails` and deprecate `Subscription` in `Transaction`
 
 ## 2.78.0
 * Add support for `OAUTH_ACCESS_REVOKED` in `WebhookNotification`s